Don't lose newline data if it's inside an otherwise empty element
Our RemoveEmptyElements method was removing nodes that had a bunch of empty <br>
tags, which meant that newlines were being stripped.

Sort of an edge case that you'd only hit if you entered in text that didn't have
a wrapping paragraph.

I'm not even sure how this is possible as we always ensure a paragraph when
values are inserted into the textara.