Ensure paragraph isn't deleted when cmd-a + text
We had code in place where if someone deleted all the content using a backspace
key, a paragraph would be regenerated, but in the situation where a user hit
cmd-a and then typed a single letter, all content was getting deleted, and no
new wrapping paragraph inserted.

This was causing problems later down the line in situations where the composer
expected to find a paragraph for things like inserting pasted content, and there
was no paragraph to grab onto.